refactor(command): Refactor config command into multiple files
